move document types out into separate files
