+++ cobalt/__init__.py
@@ -1,4 +1,5 @@
-from .act import Act, AmendmentEvent, RepealEvent
+from .hierarchical import Act, AmendmentEvent, RepealEvent
 from .uri import FrbrUri
 
+# TODO: add other document types
 __all__ = [Act, FrbrUri, AmendmentEvent, RepealEvent]
