+++ cobalt/collection.py
@@ -1,0 +1,8 @@
+from .akn import StructuredDocument
+
+
+class CollectionStructure(StructuredDocument):
+    structure_type = "collectionStructure"
+    main_content_tag = "collectionBody"
+
+
