+++ cobalt/judgment.py
@@ -1,4 +1,9 @@
-from .akn import JudgmentStructure
+from .akn import StructuredDocument
+
+
+class JudgmentStructure(StructuredDocument):
+    structure_type = "judgmentStructure"
+    main_content_tag = "judgmentBody"
 
 
 class Judgment(JudgmentStructure):
